Fixed Union and Added Intersection Fix
Union changed the bag, who would have guessed. Fixed it now. It adds the contents of both bags into a new copy (in parallel). The execution time is reduced this way. Intersection is changing the contents of the bag it is receiving, I added a fix but increases execution time (but gets the right results and doesn't change any bag contents)